+++ CHANGELOG.md
@@ -12,6 +12,8 @@ Remember to align the itemized text with the first line of an item within a list
   * JAX now requires ml_dtypes version 0.4.0 or newer.
 
 * Deprecations
+  * `jax.sharding.XLACompatibleSharding` is deprecated. Please use
+    `jax.sharding.Sharding`.
   * Removed a number of previously-deprecated APIs:
     * from {mod}`jax.core`: `non_negative_dim`, `DimSize`, `Shape`
     * from {mod}`jax.lax`: `tie_in`